+++ app/user/structs/list.go
@@ -8,7 +8,7 @@ ListUser is the API response payload for returning multiple users.
 type ListUser struct {
 	BirthDate  time.Time `json:"birthDate" binding:"required"`
 	Gender     Gender    `json:"gender" binding:"required"`
-	ID         int       `json:"id" binding:"required"`
+	ID         uint64    `json:"id" binding:"required"`
 	Name       string    `json:"name" binding:"required"`
 	Profession string    `json:"profession" binding:"required"`
 }
